update placeholder hook
