+++ src/main/scala/io/izzel/taboolib/common/plugin/bridge/BridgeProxy.java
@@ -1,0 +1,14 @@
+package io.izzel.taboolib.common.plugin.bridge;
+
+/**
+ * TabooLib
+ * io.izzel.taboolib.common.plugin.bridge.BridgeProxy
+ *
+ * @author sky
+ * @since 2021/1/11 11:10 上午
+ */
+public interface BridgeProxy {
+
+    void initProxy(String id, Object source);
+
+}
